Repair OSX static builds
Static and gitian builds fail due to the inclusion of non statically compiled, OS dependant, dependancies being included in the static build process.  This otools append is not required in a dmg build so long as the depends folder (aka static build or gitian build) process is followed, which is the reccommended method to generate reliable binaries.